% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunum.R
\docType{package}
\name{immunum-package}
\alias{immunum}
\alias{immunum-package}
\title{immunum: Fast Antibody and T-Cell Receptor Sequence Numbering}
\description{
R bindings for the \code{immunum} Rust crate. High-performance numbering
of antibody and T-cell receptor variable-domain sequences using IMGT
and Kabat schemes.
}
\section{Main entry points}{

\itemize{
\item \link{Annotator}: R6 class for numbering and segmenting sequences.
\item \code{\link[=immunum_version]{immunum_version()}}: Version of the linked Rust core.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ENPICOM/immunum}
  \item \url{https://immunum.enpicom.com}
  \item Report bugs at \url{https://github.com/ENPICOM/immunum/issues}
}

}
\author{
\strong{Maintainer}: Egor Marin \email{e.marin@enpicom.com}

Authors:
\itemize{
  \item Eli Eydlin \email{ilyabeydlin@gmail.com}
}

Other contributors:
\itemize{
  \item ENPICOM \email{dev@enpicom.com} [copyright holder]
}

}
\keyword{internal}
