% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{immunum_version}
\alias{immunum_version}
\title{Return the version of the underlying \code{immunum} Rust crate}
\usage{
immunum_version()
}
\value{
A length-1 character vector with the crate version.
}
\description{
This is the version of the Rust core, not the R package wrapper.
Returned as a \code{character(1)} to mirror Python's
\code{immunum._internal.__version__}.
}
\examples{
immunum_version()
}
